% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_type.R
\name{ae_id}
\alias{ae_id}
\alias{ae_label}
\alias{ae_condition}
\title{Get properties of an \code{ae_type} object}
\usage{
ae_id(ae_type)

ae_label(ae_type)

ae_condition(ae_type)
}
\arguments{
\item{ae_type}{An \code{ae_type} object.}
}
\description{
\code{ae_id()} and \code{ae_label()} return the \code{id} and \code{label}
properties of an \code{ae_type} object.

\code{ae_condition()} returns the stored condition. If the
condition is \code{NULL}, it returns \code{quo(TRUE)} so it can
be safely used inside \code{dplyr::filter()}.
}
\examples{
rel_ae <- make_ae_type("rel_ae", "Related AEs", AEREL == "Y")

ae_id(rel_ae)
ae_label(rel_ae)
ae_condition(rel_ae)

all_ae <- make_ae_type("all_ae", "All AEs")
ae_condition(all_ae)
}
