% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_by_aetype.R
\name{apply_ae_type}
\alias{apply_ae_type}
\title{Helper function to apply a single ae_type to a dataset (filter + add id/label)}
\usage{
apply_ae_type(data, ae_type, label_col = "label")
}
\arguments{
\item{data}{A data.frame to be filtered (e.g. ADAE).}

\item{ae_type}{An ae_type object created by make_ae_type().}

\item{label_col}{A character string specifying the name of the label column (default: "label").}
}
\value{
A data.frame with added columns \code{id}, \code{label} (or custom label column).
}
\description{
Helper function to apply a single ae_type to a dataset (filter + add id/label)
}
\keyword{internal}
