% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpn_query_ptsummary.R
\name{jpn_query_ptsummary}
\alias{jpn_query_ptsummary}
\title{Create ARD for adverse event overview (by AETYPE and PT(AEDECOD))}
\usage{
jpn_query_ptsummary(data, denominator, ae_types, by = NULL)
}
\arguments{
\item{data}{A data.frame (ADAE-like). Unlike \code{\link[=jpn_query_base]{jpn_query_base()}}, \code{data} does not
need to contain \code{AETYPE} in advance.}

\item{denominator}{A data.frame used as the denominator in \code{{cards}}.
Must contain the columns specified in \code{by} (if provided).}

\item{ae_types}{List of \code{ae_type} objects passed to \code{bind_by_aetype()} to create \code{AETYPE}.}

\item{by}{Optional tidyselect specification of column(s) used for column splits
(e.g. treatment arm, subgroup). If \code{NULL} (default), no \code{by} columns are used.}
}
\value{
A \code{{cards}} ARD object returned by \code{cards::ard_stack_hierarchical()}.
}
\description{
Convenience wrapper around \code{\link[=jpn_query_aggregate]{jpn_query_aggregate()}} that summarizes data
Aggregate by AETYPE and AEDECOD.
}
\examples{
ae_types <- list(
  make_ae_type("any_ae", "Any AE"),
  make_ae_type("rel_ae", "Related AEs", AEREL == "Y")
)

# AETYPE x PT (AEDECOD) summary, split by treatment arm
jpn_query_ptsummary(
  data        = adae,
  denominator = adsl,
  ae_types    = ae_types,
  by          = TRT01A
)

}
