% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_type.R
\name{make_ae_type}
\alias{make_ae_type}
\title{Create an \code{ae_type} object}
\usage{
make_ae_type(id, label, condition = NULL)
}
\arguments{
\item{id}{ID value for the object.}

\item{label}{Label for the AE type.}

\item{condition}{Expression for subsetting the dataset.}
}
\value{
An \code{ae_type} object.
}
\description{
Create an \code{ae_type} object
}
\examples{
make_ae_type("all_ae", "All AEs")
make_ae_type("rel_ae", "Related AEs", AEREL == "Y")
}
