% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_type.R
\name{merge_ae_types}
\alias{merge_ae_types}
\title{Combine multiple \code{ae_type} objects}
\usage{
merge_ae_types(ae_types, id, label, op = c("&", "|"))
}
\arguments{
\item{ae_types}{List of \code{ae_type} objects.}

\item{id}{ID for the combined object.}

\item{label}{Label for the combined object.}

\item{op}{Logical operator used to combine conditions (\code{&} or \code{|}).}
}
\value{
An \code{ae_type} object.
}
\description{
Combine multiple \code{ae_type} objects
}
\examples{
rel_ae <- make_ae_type("rel_ae", "Related AEs", AEREL == "Y")
ser_ae <- make_ae_type("ser_ae", "Serious AEs", AESER == "Y")

merge_ae_types(
  list(rel_ae, ser_ae),
  id = "rel_ser_ae",
  label = "Serious Related AEs"
)
}
